+++ mayan/apps/common/links.py
@@ -9,7 +9,6 @@ from .icons import (
     icon_forum, icon_license, icon_setup, icon_source_code, icon_support,
     icon_tools
 )
-from .permissions_runtime import permission_error_log_view
 
 
 def get_kwargs_factory(variable_name):
@@ -50,18 +49,6 @@ link_current_user_locale_profile_edit = Link(
 link_documentation = Link(
     icon_class=icon_documentation, tags='new_window',
     text=_('Documentation'), url='https://docs.mayan-edms.com'
-)
-link_object_error_list = Link(
-    icon_class_path='mayan.apps.common.icons.icon_object_error_list',
-    kwargs=get_kwargs_factory('resolved_object'),
-    permissions=(permission_error_log_view,), text=_('Errors'),
-    view='common:object_error_list',
-)
-link_object_error_list_clear = Link(
-    icon_class_path='mayan.apps.common.icons.icon_object_error_list_clear',
-    kwargs=get_kwargs_factory('resolved_object'),
-    permissions=(permission_error_log_view,), text=_('Clear all'),
-    view='common:object_error_list_clear',
 )
 link_forum = Link(
     icon_class=icon_forum, tags='new_window', text=_('Forum'),