+++ mayan/apps/common/models.py
@@ -3,46 +3,16 @@ import uuid
 from pytz import common_timezones
 
 from django.conf import settings
-from django.contrib.contenttypes.fields import GenericForeignKey
-from django.contrib.contenttypes.models import ContentType
 from django.db import models
 from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 
-from .managers import ErrorLogEntryManager, UserLocaleProfileManager
+from .managers import UserLocaleProfileManager
 
 
+# Required by migration 0010_auto_20180403_0702.py
 def upload_to(instance, filename):
     return 'shared-file-{}'.format(uuid.uuid4().hex)
-
-
-class ErrorLogEntry(models.Model):
-    """
-    Class to store an error log for any object. Uses generic foreign keys to
-    reference the parent object.
-    """
-    namespace = models.CharField(
-        max_length=128, verbose_name=_('Namespace')
-    )
-    content_type = models.ForeignKey(
-        blank=True, on_delete=models.CASCADE, null=True,
-        related_name='error_log_content_type', to=ContentType,
-    )
-    object_id = models.PositiveIntegerField(blank=True, null=True)
-    content_object = GenericForeignKey(
-        ct_field='content_type', fk_field='object_id',
-    )
-    datetime = models.DateTimeField(
-        auto_now_add=True, db_index=True, verbose_name=_('Date time')
-    )
-    result = models.TextField(blank=True, null=True, verbose_name=_('Result'))
-
-    objects = ErrorLogEntryManager()
-
-    class Meta:
-        ordering = ('datetime',)
-        verbose_name = _('Error log entry')
-        verbose_name_plural = _('Error log entries')
 
 
 class UserLocaleProfile(models.Model):